requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# server.py: 3
Flask == 0.12.2

# server.py: 6
PyYAML == 3.12

# server.py: 8
click == 6.7
